refactor: only allow shared lanugages
