+++ src/components/landing/home.js
