+++ src/components/landing/internet-economy.js
