+++ src/components/landing/latest.js
@@ -2,7 +2,7 @@ import { Col, Row } from 'antd'
 import React, { useState, useEffect } from 'react'
 import styled from 'styled-components'
 import text from './translations'
-import NewsBox from '../../components/news-box'
+import NewsBox from '../news-box'
 import Twitter from '../../../static/assets/images/twitter.png'
 import Github from '../../../static/assets/images/github.png'
 import Forum from '../../../static/assets/images/forum.png'
