+++ src/components/landing/partners.js
