+++ src/components/landing/translations.js
